frontend: Replace GameView
